Trying to speed up code.

+++ hazma/decay_helper_functions/decay_charged_pion.pxd
@@ -3,14 +3,17 @@ import numpy as np
 cimport numpy as np
 
 cdef class ChargedPion:
-    cdef float __gamma(self, float eng, float mass)
-    cdef float __beta(self, float eng, float mass)
-    cdef float __eng_gam_max(self, float engPi)
-    cdef float __integrand(self, float cl, float engGam, float engPi)
-    cdef float __eng_gam_max_mu_rf
-    cdef float __eng_mu_pi_rf
-    cdef float __muon_spectrum(self, float eng_gam)
+    cdef double __gamma(self, double eng, double mass)
+    cdef double __beta(self, double eng, double mass)
+    cdef double __eng_gam_max(self, double engPi)
+    cdef double __integrand(self, double cl, double engGam, double engPi)
+    cdef double __eng_gam_max_mu_rf
+    cdef double __eng_mu_pi_rf
+    cdef double __muon_spectrum(self, double eng_gam)
 
     cdef decay_muon.Muon __muon
     cdef np.ndarray __eng_gams_mu
     cdef np.ndarray __mu_spec
+    #cpdef SpectrumPoint(self, double eng_gam, double eng_pi)
+    #cpdef Spectrum(self, np.ndarray[np.float64_t, ndim=1] eng_gams, \
+    #               double eng_pi)
